Fuse additional optimizations
